add right click wiki to GE plugin search
